% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-and-end-dates.R
\name{bindates}
\alias{bindates}
\title{Bin Dates}
\usage{
bindates(x, orders, sincewhen = Sys.Date(), neverfill = "never")
}
\arguments{
\item{x}{a character vector of date-type data}

\item{orders}{a charater vector of posssible formats}

\item{sincewhen}{a cut off date to compare x to}

\item{neverfill}{if x is NA, what value should be used}
}
\value{
a factor vector with values of "this yr", "last 3 yrs", "last 5 yrs, "last 10 yrs", "long ago" and "never (or whatever you pick for neverfill)
}
\description{
parse and seperate dates into bins
}
