% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-AF-models.R
\name{buildAFmodels}
\alias{buildAFmodels}
\title{Build AF Models}
\usage{
buildAFmodels(trainingdata, models = c("rf", "lm", "glm"))
}
\arguments{
\item{trainingdata}{a data frame of training data, probably built by buildAFpredictors().}

\item{models}{a vector of models to build. Options include rf (random forest), lm (linear regression) and glm (generalize linear regression).}
}
\value{
a list of models. The names of the models will be be xxmodel, based on the options included in the models argument.
}
\description{
Build models to predict annual fund giving.
}

