% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAFpredictors.R
\name{buildAFpredictors}
\alias{buildAFpredictors}
\title{Build AF Predictors}
\usage{
buildAFpredictors(trainingyear, yeartype = "fiscal",
  trainingsource = hallptbl, primaryonly = T)
}
\arguments{
\item{trainingyear}{An integer value of the year in 4 digit format (eg. 2012) of the data that should be used to train the models.}

\item{yeartype}{A character string describing the type of year. Choices should be "fiscal" or "calendar".}

\item{trainingsource}{A hallp-style source for the training data. Valid choises include database tables or a csv location.}
}
\value{
A data frame of data ready with pidms, 3 outcome variables (prefixed with "outcome_") and 42 other variables. Note that variables DO include previous years' giving.
}
\description{
Builds a data frame suitable for doing prediction for annual fund.
}

