% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-demographic-data.R
\name{builddemographicdata}
\alias{builddemographicdata}
\title{Build Demographic Data}
\usage{
builddemographicdata(trainingdata, startdate, primaryonly = T)
}
\arguments{
\item{trainingdata}{a hallp-style data frame.}

\item{primaryonly}{a logical value indicating whether the final result should include only primary people.}
}
\value{
a data frame with the established predictor data.
}
\description{
build a set of predictors from a set of hallp data
}

