% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-and-end-dates.R
\name{createAFcampaign}
\alias{createAFcampaign}
\title{Create AF Campaign}
\usage{
createAFcampaign(trainingyear, yeartype)
}
\arguments{
\item{trainingyear}{a integer describing the year}

\item{yeartype}{a character string describing the type of year, either "fiscal" or "calendar".}
}
\value{
a character string of the AF campaign within that year. If yeartype is not "fiscal", returns "AF%" (so other functions match any AF gift within the date range).
}
\description{
determine the AF campaign based on a year
}

