% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-and-end-dates.R
\name{createenddate}
\alias{createenddate}
\title{Create End Date}
\usage{
createenddate(trainingyear, yeartype)
}
\arguments{
\item{trainingyear}{a integer describing the year}

\item{yeartype}{a character string describing the type of year, either "fiscal" or "calendar".}
}
\value{
the POSIXct date on which trainingyear ended
}
\description{
create a POSIXct date of when the year ends
}
