% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-and-end-dates.R
\name{createstartdate}
\alias{createstartdate}
\title{Create Start Date}
\usage{
createstartdate(trainingyear, yeartype)
}
\arguments{
\item{trainingyear}{a integer describing the year}

\item{yeartype}{a character string describing the type of year, either "fiscal" or "calendar".}
}
\value{
the POSIXct date on which trainingyear started
}
\description{
create a POSIXct date of when the year starts
}
