% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-presidents.R
\name{findPresidents}
\alias{findPresidents}
\title{Find Presidents and CEOs}
\usage{
findPresidents(x)
}
\arguments{
\item{x}{a character vector of job titles}
}
\value{
a logical vector indicating whether the job title is any of the following: president, CEO, chairman, owner or self-employed. Vice presidents return FALSE.
}
\description{
convert a list of character vectors into logical based on whether each title is a top-level executive or not. Designed for use within dplyr::mutate.
}
