% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-catg-model.R
\name{measure_catg_model}
\alias{measure_catg_model}
\title{Measure the Performance of a Categorical Model}
\usage{
measure_catg_model(model, testdata, depvar)
}
\arguments{
\item{model}{a predictive model which predicts categories. Probably ought to be a caret model for now.}

\item{testdata}{the data on which you which to test and predict}

\item{depvar}{a vector of the dependent variable. Must be a factor with the same levels as what the model will predict.}
}
\value{
a one-row data frame with data about the model's accuracy including precision/recall stats and the Area Under the PR Curve.
}
\description{
return a single row of descriptive statistics making it easier to asses models
}
