% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-and-end-dates.R
\name{parsedates}
\alias{parsedates}
\title{Parse Dates}
\usage{
parsedates(x, orders, fill = NA)
}
\arguments{
\item{x}{a vector of date-time type data to parse}

\item{orders}{a character vector of possible date formats}

\item{fill}{value to fill in known NA values.}
}
\value{
a POSIXct vector
}
\description{
combines lubridate's parse_date_time and guess_formats functions to parse a variety of date formats
}
