% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-AF.R
\name{predictAF}
\alias{predictAF}
\title{Predict AF}
\usage{
predictAF(models, source, buildsource = T, currentyear, yeartype = "fiscal",
  pidms = NA)
}
\arguments{
\item{models}{a list of models to predict from. Probably built with buildAFmodels().}

\item{source}{a hallp-like source of data to do predictions from. Valid options include database tables and csvs.}

\item{buildsource}{should the prediction data be built from source. If F, source is passed directly to train() without being processed by buildAFpredictors().}

\item{currentyear}{an integer describing the current year.}

\item{yeartype}{a character string describing the type of currentyear, either 'fiscal' or 'calendar'.}

\item{pidms}{a vector of pidms to do predictions on. If left NA, predictAF() will build predictions for everyone in source.}
}
\value{
a data frame of pidms and one variable for each model used.
}
\description{
generate predictions from models with regard to AF giving.
}

