% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-training-data.R
\name{readtrainingdata}
\alias{readtrainingdata}
\title{Read Training Data}
\usage{
readtrainingdata(source)
}
\arguments{
\item{source}{A source for the data. Can be either a hallp-style database table or a csv.}
}
\value{
a hallp-style data frame with the fields necessary to build the training and prediction data.
}
\description{
Build a hallp-style data frame of raw data from a source.
}

